Fix missing methods by using files from the jar

+++ src/main/java/fr/synchroneyes/mineral/Utils/Radius.java
@@ -3,45 +3,12 @@ package fr.synchroneyes.mineral.Utils;
 import org.bukkit.Location;
 
 public class Radius {
-
     public static boolean isBlockInRadius(Location source, Location blockToFind, int radius) {
-        return (
-                (
-                        (
-                                source.getX() - radius < blockToFind.getX() &&
-                                        source.getX() + radius > blockToFind.getX()
-                        ) &&
-                                (
-                                        source.getZ() - radius < blockToFind.getZ() &&
-                                                source.getZ() + radius > blockToFind.getZ()
-                                ) &&
-                                (
-                                        source.getY() - radius < blockToFind.getY() &&
-                                                source.getY() + radius > blockToFind.getY()
-                                )
-                )
-        );
-
+        return source.getX() - (double)radius < blockToFind.getX() && source.getX() + (double)radius > blockToFind.getX() && source.getZ() - (double)radius < blockToFind.getZ() && source.getZ() + (double)radius > blockToFind.getZ() && source.getY() - (double)radius < blockToFind.getY() && source.getY() + (double)radius > blockToFind.getY();
     }
 
     public static boolean isBlockInRadiusWithDividedYAxis(Location source, Location blockToFind, int radius, int divider) {
-
-        return (
-                (
-                        (
-                                source.getX() - radius < blockToFind.getX() &&
-                                        source.getX() + radius > blockToFind.getX()
-                        ) &&
-                                (
-                                        source.getZ() - radius < blockToFind.getZ() &&
-                                                source.getZ() + radius > blockToFind.getZ()
-                                ) &&
-                                (
-                                        source.getY() - (int) (radius / divider) < blockToFind.getY() &&
-                                                source.getY() + (int) (radius / divider) > blockToFind.getY()
-                                )
-                )
-        );
-
+        return source.getX() - (double)radius < blockToFind.getX() && source.getX() + (double)radius > blockToFind.getX() && source.getZ() - (double)radius < blockToFind.getZ() && source.getZ() + (double)radius > blockToFind.getZ() && source.getY() - (double)(radius / divider) < blockToFind.getY() && source.getY() + (double)(radius / divider) > blockToFind.getY();
     }
 }
+
